Bark at each team in the config if none provided
Context
=======

I'm lazy.

Much like I don't want to add and remove people as they join and leave a
project (see
https://github.com/envato/seal/commit/0b59c77781b79bd06d38dc54898266892d0e85eb),
I don't want to have to maintain Heroku schedulers (as neat as they are)
each time somebody adds or removes a team.

Change
======

If a team is not specified, Seal will bark at all teams in its config.
If a team is provided, Seal will just bark at that one, as before.