Use github search to power pull request search
Currently, the seal loops through all repos and fetches all pull
requests. We could replace this with the issue search:
https://developer.github.com/v3/search/#search-issues

This will return all open PRs in all alphagov repos. It uses the same
search syntax as github.com.

This means 1) we can get rid of the repos in the configuration, 2) new
repos are added automatically, 3) we save a bunch of requests.

Closes: #52