Let seal sea more pull requests
It turns out `#search_issues` returns 30 results per default. This
caused the seal to report on only the 30 most recent pull requests in
the organisation.

This commit fixes this by increasing that number to 100, the maximum of
requests per page according to the API docs (and experimentation). This
should take care of most missing pull requests.