Refactor seal bash scripts
Bash has arrays which you can use to iterate instead of creating several
repeated lines.

Also since you not in ruby but in bash, you can dismmiss the double '`'
characters around the command you want to execute.

They are normally used when you need the output, for example

```
vim `which ruby`

vim $(which ruby)
```

My suggestion would be for you to play around with ARGV and make
`./bin/seal.rb` receive a list of teams and the given command (`quotes`
in your case) and iterate those in ruby, maybe something for you to
think about...

Thanks for taking time to maintain seal :)