Move Chris & Tatiana to Custom
Whilst they are there as part of the mainstream pop-up team.

I wasn't sure maybe you needed to split custom into some people that
would use the #custom channel and some that would use
the #mainstream-pop-up channel. That seemed a bit more destructive
though so let me know if it should be edited to that.