Eliminate Jen's vestigial remains
When [Jen regenerated into Shilpa][1], we experienced a rift in the space-time
continuum resulting in them both having corporeal representations in the same
plane of existence (the seal).  This must be resolved prior to the annual
multiverse reconciliation date or we may all be annihilated.

There can be only one.

[1]: https://github.com/binaryberry/seal/commit/60500a4f9f7fe7184cf4ecb5d5a729299b856dcf