Add Content Tools
The Content Tools team now look after a bunch of apps that many others
issue PRs on. We need to be more aware of these and have Seal
prompt us.

I’ve added the team to the `config.yml` with our current members,
the labels I think we’d want to ignore (Do not merge) and have
stolen the Core Format team’s charter for quotes.