Show check mark when PRs are approved.
Many teams are now using Github's new review workflow, where you
explicitly approve a PR rather than marking it with a +1. This is now
supported in their API as a preview function via a specific media type
header.

This change adds a green check mark for PRs that are approved in this
way.