Remove draft PRs from seal display
Given the purpose of the seal is to remind people to take action towards
getting PRs reviewed and then merged, it doesn't particularly make sense
for these to include draft PRs. As the status of draft indicates a PR is
not ready to be merged.

Documentation to show drafts:
https://help.github.com/en/articles/searching-issues-and-pull-requests#search-for-draft-pull-requests
and exclude certain results:
https://help.github.com/en/articles/understanding-the-search-syntax#exclude-certain-results